refactor: Changing monolit to MS Application
