+++ key-generator-service/src/main/java/com/isaacandrade/keygeneratorservice/KeyGeneratorServiceApplication.java
@@ -1,0 +1,13 @@
+package com.isaacandrade.keygeneratorservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KeyGeneratorServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(KeyGeneratorServiceApplication.class, args);
+    }
+
+}
