+++ key-generator-service/src/test/java/com/isaacandrade/keygeneratorservice/KeyGeneratorServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.isaacandrade.keygeneratorservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class KeyGeneratorServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
