+++ resolver-service/src/main/java/com/isaacandrade/resolverservice/ResolverServiceApplication.java
@@ -1,0 +1,13 @@
+package com.isaacandrade.resolverservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ResolverServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ResolverServiceApplication.class, args);
+	}
+
+}
