+++ resolver-service/src/test/java/com/isaacandrade/resolverservice/ResolverServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.isaacandrade.resolverservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ResolverServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
