+++ shortener-service/src/main/java/com/isaacandrade/url_shortener_service/UrlShortenerApplication.java
@@ -1,4 +1,4 @@
-package com.isaacandrade.url_shortener;
+package com.isaacandrade.url_shortener_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
