+++ shortener-service/src/main/resources/application.yml
@@ -17,7 +17,6 @@ spring:
       host: ${REDIS_HOST}
       port: ${REDIS_PORT}
       password: ${REDIS_PASSWORD}
-      timeout: 5000
       client-type: lettuce
   cache:
     type: ${SPRING_CACHE_TYPE}
