--- src/main/java/com/isaacandrade/url_shortener/domain/TestEntity.java
@@ -1,54 +1,0 @@
-package com.isaacandrade.url_shortener.domain;
-
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.util.Objects;
-
-@Document(collection = "test_collection")
-public class TestEntity {
-    @Id
-    private String id;
-    private String message;
-
-    public TestEntity(String id, String message) {
-        this.id = id;
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    @Override
-    public String toString() {
-        return "TestEntity{" +
-                "id='" + id + '\'' +
-                ", message='" + message + '\'' +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (o == null || getClass() != o.getClass()) return false;
-        TestEntity that = (TestEntity) o;
-        return Objects.equals(id, that.id) && Objects.equals(message, that.message);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, message);
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
